Schemas are now lazy
Before this patch, schemas were computed as soon as a RPC was registered.
With this patch, schemas are lazy, and they are only computed if they
are actually used. In practice, we don't need them that often, and they
can significantly slow down registration.

This is a breaking change as it changes the interface: you may need to
insert some Lazy.force or Lazy.from_val.