fix opam dependencies
ocamformat: Adding ocamlformat as dependencies of an opam package
seems to affect `resto` packages when pinned locally. This commit
removes ocamlformat as dependency from resto and instead adds it
to .ocamlformat file.

Additionally, we explicitly state dependencies in both dune and
opam files so that 'opam pin . ' command works correctly.

Lastly, the repo is now formatted with ocamlformat.0.19.0.

Signed-off-by: Bikal Lem <gbikal+git@gmail.com>